Revert "Checking short commit over Eclipse"
This reverts commit 17fdbe22b4750da86b694863403e7f624049c820.

+++ Source/OurFood/src/main/java/ourfood/service/CropService.java
@@ -6,7 +6,7 @@ import ourfood.domain.Crop;
 
 public interface CropService {
 
-    Crop getCrop(int id);
+    Crop getCrop(Long id);
 
     void saveCrop(Crop crop);
